requirements.txt
pip install SentencePiece sklearn matplotlib sentence-transformers jupyter
tokenizers-0.13.3-cp38-cp38-manylinux_2_17_x86_64.manylinux2014_x86_64.whl
torch-1.8.2+cu111-cp38-cp38-linux_x86_64.whl
torchtext-0.9.2-cp38-cp38-linux_x86_64.whl
torchvision-0.9.2+cu111-cp38-cp38-linux_x86_64.whl
transformers-4.28.1-py3-none-any.whl

